requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
git+https://github.com/facebookresearch/sam2.git
streamlit-plotly-events>=0.0.6
plotly>=5.0.0
